requirements.txt
Flask==2.0.3
itsdangerous==2.1.0
gunicorn==20.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
Werkzeug==20.1.0
requests==2.22.0
numpy==1.21.2
scikit-learn==1.0.2
Pillow==9.2.0
tensorflow-cpu==2.8.0
